create new PriceForm
